refactor: 에이전트 반환 형식 공통화
