+++ src/main/java/datastreams_knu/bigpicture/interest/agent/InterestCrawlingAssistant.java
@@ -1,6 +1,6 @@
 package datastreams_knu.bigpicture.interest.agent;
 
-import datastreams_knu.bigpicture.interest.agent.dto.InterestCrawlingResultDto;
+import datastreams_knu.bigpicture.common.dto.CrawlingResultDto;
 import dev.langchain4j.service.SystemMessage;
 import dev.langchain4j.service.UserMessage;
 import dev.langchain4j.service.V;
@@ -26,5 +26,5 @@ public interface InterestCrawlingAssistant {
             정확하고 유용한 데이터만 제공하세요.
         """)
     @UserMessage("type: {{type}}, n: {{n}}")
-    InterestCrawlingResultDto execute(@V("type") String type, @V("n") int n);
+    CrawlingResultDto execute(@V("type") String type, @V("n") int n);
 }
