+++ src/main/java/datastreams_knu/bigpicture/interest/repository/KoreaInterestRepository.java
@@ -1,6 +1,6 @@
 package datastreams_knu.bigpicture.interest.repository;
 
-import datastreams_knu.bigpicture.interest.domain.KoreaInterest;
+import datastreams_knu.bigpicture.interest.entity.KoreaInterest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface KoreaInterestRepository extends JpaRepository<KoreaInterest, Long> {
