+++ src/main/java/datastreams_knu/bigpicture/interest/repository/USInterestRepository.java
@@ -1,6 +1,6 @@
 package datastreams_knu.bigpicture.interest.repository;
 
-import datastreams_knu.bigpicture.interest.domain.USInterest;
+import datastreams_knu.bigpicture.interest.entity.USInterest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface USInterestRepository extends JpaRepository<USInterest, Long> {
